refactor: rename hook_input.py → session_start_input.py
The file contains SessionStartHookInput — the name should reflect that
it's specifically for session start hook input, not hooks in general.
