+++ devinfra/claude/claude_api/post_tool_use.py
@@ -12,7 +12,7 @@ from typing import Any, Literal
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 from pydantic.alias_generators import to_camel
 
-from devinfra.claude.claude_api.hook_input import PermissionMode
+from devinfra.claude.claude_api.session_start_input import PermissionMode
 
 
 class _CamelModel(BaseModel):
