+++ devinfra/claude/session_start.py
@@ -36,7 +36,7 @@ from devinfra.claude import (
     tmpfs_setup,
 )
 from devinfra.claude.auth_proxy import setup as proxy_setup
-from devinfra.claude.claude_api.hook_input import SessionStartHookInput
+from devinfra.claude.claude_api.session_start_input import SessionStartHookInput
 from devinfra.claude.debug import log_entrypoint_debug
 from devinfra.claude.errors import SkipError
 from devinfra.claude.hook_config import OtelConfig
